refactor: move parsers to internal packages
- Create internal/types for shared types (Line, SupportedTypes)
- Create internal/common for shared functions (SanitizeAsm, ParseObjectDump)
- Create internal/amd64, internal/arm64, internal/loong64, internal/riscv64
- Each architecture parser unified as parser.go
- Variables prefixed with architecture to avoid symbol conflicts
- Update main.go to use types and common packages
- Delete parser_common.go (moved to internal/common)
- Fix compilation errors and ensure build succeeds
