+++ parser_select.go
@@ -6,25 +6,33 @@
 //
 // http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by local law or agreed to in writing, software
+// Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 package main
 
-import "fmt"
+import (
+	"fmt"
 
-func parseAssemblyForTarget(target, path string) (map[string][]Line, map[string]int, error) {
+	"github.com/gorse-io/goat/internal/amd64"
+	"github.com/gorse-io/goat/internal/arm64"
+	"github.com/gorse-io/goat/internal/loong64"
+	"github.com/gorse-io/goat/internal/riscv64"
+	"github.com/gorse-io/goat/internal/types"
+)
+
+func parseAssemblyForTarget(target, path string) (map[string][]types.Line, map[string]int, error) {
 	switch target {
 	case "amd64":
-		return parseAssemblyX86(path)
+		return amd64.ParseAssembly(path)
 	case "arm64":
-		return parseAssemblyA64(path)
+		return arm64.ParseAssembly(path)
 	case "loong64":
-		return parseAssemblyLoong(path)
+		return loong64.ParseAssembly(path)
 	case "riscv64":
-		return parseAssemblyRv64(path)
+		return riscv64.ParseAssembly(path)
 	default:
 		return nil, nil, fmt.Errorf("unsupported target architecture %q", target)
 	}
